# contracts/openapi/pipe4-v1.json
{
  "components": {
    "schemas": {
...
      "ApiErrorBody": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ApiErrorBody",
        "type": "object"
      },
      "ApiFailure": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorBody"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          },
          "success": {
            "default": false,
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "error",
          "meta"
        ],
        "title": "ApiFailure",
        "type": "object"
      },
      "ApiMeta": {
        "additionalProperties": false,
        "properties": {
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "timestamp"
        ],
        "title": "ApiMeta",
        "type": "object"
      },
...
      "ApiSuccess_PolicyInspection_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PolicyInspection"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          },
          "success": {
            "default": true,
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "ApiSuccess[PolicyInspection]",
        "type": "object"
      },
...
      "ApiSuccess_SourceProfile_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/SourceProfile"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          },
          "success": {
            "default": true,
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "ApiSuccess[SourceProfile]",
        "type": "object"
      },
...
      "ApiSuccess_StateVersion_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/StateVersion"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          },
          "success": {
            "default": true,
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "ApiSuccess[StateVersion]",
        "type": "object"
      },
...
      "ApiSuccess_Union_StateVersion__NoneType__": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StateVersion"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          },
          "success": {
            "default": true,
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "ApiSuccess[Union[StateVersion, NoneType]]",
        "type": "object"
      },
...
      "ApiSuccess_list_DomainEvent__": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/DomainEvent"
            },
            "title": "Data",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          },
          "success": {
            "default": true,
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "ApiSuccess[list[DomainEvent]]",
        "type": "object"
      },
...
      "DomainEvent": {
        "additionalProperties": false,
        "properties": {
          "aggregate_id": {
            "title": "Aggregate Id",
            "type": "string"
          },
          "aggregate_type": {
            "title": "Aggregate Type",
            "type": "string"
          },
          "correlation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correlation Id"
          },
          "event_id": {
            "title": "Event Id",
            "type": "string"
          },
          "event_type": {
            "title": "Event Type",
            "type": "string"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "payload": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "number"
                },
                {
                  "type": "boolean"
                },
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Payload",
            "type": "object"
          }
        },
        "required": [
          "event_type",
          "aggregate_type",
          "aggregate_id",
          "occurred_at"
        ],
        "title": "DomainEvent",
        "type": "object"
      },
...
      "PolicyInspection": {
        "additionalProperties": false,
        "properties": {
          "content_hash": {
            "title": "Content Hash",
            "type": "string"
          },
          "decision_contexts": {
            "items": {
              "type": "string"
            },
            "title": "Decision Contexts",
            "type": "array"
          },
          "nearby_weights": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Nearby Weights",
            "type": "object"
          },
          "state_types": {
            "items": {
              "type": "string"
            },
            "title": "State Types",
            "type": "array"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "version",
          "content_hash",
          "state_types",
          "decision_contexts",
          "nearby_weights"
        ],
        "title": "PolicyInspection",
        "type": "object"
      },
...
      "SourceCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "authority_scopes": {
            "items": {
              "type": "string"
            },
            "title": "Authority Scopes",
            "type": "array",
            "uniqueItems": true
          },
          "cost_units": {
            "default": 0,
            "minimum": 0.0,
            "title": "Cost Units",
            "type": "number"
          },
          "evidence_capability": {
            "default": 0.5,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Evidence Capability",
            "type": "number"
          },
          "expected_latency_ms": {
            "default": 1000,
            "minimum": 0.0,
            "title": "Expected Latency Ms",
            "type": "integer"
          },
          "provider_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Key"
          },
          "reliability": {
            "default": 0.5,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Reliability",
            "type": "number"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          },
          "source_type": {
            "title": "Source Type",
            "type": "string"
          }
        },
        "required": [
          "source_id",
          "source_type"
        ],
        "title": "SourceCreateRequest",
        "type": "object"
      },
      "SourceOutcomeRequest": {
        "additionalProperties": false,
        "properties": {
          "correct": {
            "title": "Correct",
            "type": "boolean"
          },
          "entity_id": {
            "title": "Entity Id",
            "type": "string"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          },
          "state_type": {
            "$ref": "#/components/schemas/StateType"
          }
        },
        "required": [
          "source_id",
          "entity_id",
          "state_type",
          "correct"
        ],
        "title": "SourceOutcomeRequest",
        "type": "object"
      },
...
      "StateType": {
        "enum": [
          "AVAILABLE",
          "ACCESSIBLE",
          "WORKING",
          "TIME",
          "CHANGED"
        ],
        "title": "StateType",
        "type": "string"
      },
...
      "VerificationRequest": {
        "additionalProperties": false,
        "properties": {
          "observation_id": {
            "title": "Observation Id",
            "type": "string"
          }
        },
        "required": [
          "observation_id"
        ],
        "title": "VerificationRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-Internal-Token",
        "type": "apiKey"
      },
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Horizontal current-reality resolution, evidence, verification, WATCH, access, nearby and acquisition service.",
    "title": "UWACI Pipe 4 Current Reality API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/internal/pipe4/events": {
      "get": {
        "operationId": "events_api_v1_internal_pipe4_events_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccess_list_DomainEvent__"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Invalid request intent"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Permission denied"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "State/resource conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Validation failed"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Rate limited"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Unexpected server error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Dependency unavailable"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Events",
        "tags": [
          "Pipe 4 internal"
        ]
      }
    },
    "/api/v1/internal/pipe4/policies": {
      "get": {
        "operationId": "policies_api_v1_internal_pipe4_policies_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccess_PolicyInspection_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Invalid request intent"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Permission denied"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "State/resource conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Validation failed"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Rate limited"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Unexpected server error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Dependency unavailable"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Policies",
        "tags": [
          "Pipe 4 internal"
        ]
      }
    },
    "/api/v1/internal/pipe4/source-outcomes": {
      "post": {
        "operationId": "source_outcome_api_v1_internal_pipe4_source_outcomes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceOutcomeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccess_SourceProfile_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Invalid request intent"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Permission denied"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "State/resource conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Validation failed"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Rate limited"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Unexpected server error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Dependency unavailable"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Source Outcome",
        "tags": [
          "Pipe 4 internal"
        ]
      }
    },
    "/api/v1/internal/pipe4/sources": {
      "post": {
        "operationId": "source_create_api_v1_internal_pipe4_sources_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccess_SourceProfile_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Invalid request intent"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Permission denied"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "State/resource conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Validation failed"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Rate limited"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Unexpected server error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Dependency unavailable"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Source Create",
        "tags": [
          "Pipe 4 internal"
        ]
      }
    },
    "/api/v1/internal/pipe4/state-versions/{state_version_id}/expire": {
      "post": {
        "operationId": "expire_api_v1_internal_pipe4_state_versions__state_version_id__expire_post",
        "parameters": [
          {
            "in": "path",
            "name": "state_version_id",
            "required": true,
            "schema": {
              "title": "State Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccess_StateVersion_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Invalid request intent"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Permission denied"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "State/resource conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Validation failed"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Rate limited"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Unexpected server error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Dependency unavailable"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Expire",
        "tags": [
          "Pipe 4 internal"
        ]
      }
    },
    "/api/v1/internal/pipe4/verifications": {
      "post": {
        "operationId": "verify_api_v1_internal_pipe4_verifications_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerificationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccess_Union_StateVersion__NoneType__"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Invalid request intent"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Permission denied"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "State/resource conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Validation failed"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Rate limited"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Unexpected server error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Dependency unavailable"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Verify",
        "tags": [
          "Pipe 4 internal"
        ]
      }
    },
...
  }
}